// store/hotel/types.ts
export enum HotelActionTypes {
    GET_HOTELS = '[Hotel] GET_HOTELS',
    GET_HOTELS_SUCCESS = '[Hotel] GET_HOTELS_SUCCESS',
    GET_HOTELS_FAIL = '[Hotel] GET_HOTELS_FAIL',
    SELECT_HOTEL = '[Hotel] SELECT_HOTEL',
    SELECT_ALL_HOTELS = '[Hotel] SELECT_ALL_HOTELS',
    FILTER_HOTELS = '[Hotel] FILTER_HOTELS',
    GET_HOTEL_PLANS = '[Hotel] GET_HOTEL_PLANS',
    GET_HOTEL_PLANS_SUCCESS = '[Hotel] GET_HOTEL_PLANS_SUCCESS',
    GET_HOTEL_PLANS_FAIL = '[Hotel] GET_HOTEL_PLANS_FAIL',
    GET_ROOM_TYPES = '[Rooms] GET_ROOM_TYPES',
    GET_ROOM_TYPES_SUCCESS = '[Rooms] GET_ROOM_TYPES_SUCCESS',
    GET_ROOM_TYPES_FAIL = '[Rooms] GET_ROOM_TYPES_FAIL',
    GET_TIERS_LIST = '[Hotel] GET_TIERS_LIST',
    GET_TIERS_LIST_SUCCESS = '[Hotel] GET_TIERS_LIST',
    GET_TIERS_LIST_FAIL = '[Hotel] GET_TIERS_LIST',
    SAVE_MODEL_FILTER = '[Hotel] SAVE_MODEL_FILTER',
    SAVE_RECENT_SEARCH = '[Hotel] SAVE_RECENT_SEARCH',
    GET_HOTEL_PRICE = '[Hotel] GET_HOTEL_PRICE',
    GET_HOTEL_PRICE_SUCCESS = '[Hotel] GET_HOTEL_PRICE_SUCCESS',
    GET_HOTEL_PRICE_FAIL = '[Hotel] GET_HOTEL_PRICE_FAIL',
    SAVE_SORT_BY_VALUE = '[Hotel] SAVE_SORT_BY_VALUE',
    RESET_FILTER_MODEL = '[Hotel] RESET_FILTER_MODEL',
    SAVE_FULL_PATH = '[Hotel] SAVE_FULL_PATH',
    SAVE_FULL_PATH_HOTEL_DETAIL = '[Hotel] SAVE_FULL_PATH_HOTEL_DETAIL',
    CLEAR_TIERS_LIST = '[Hotel] CLEAR_TIERS_LIST',
    SAVE_HOTEL_ID = '[Hotel] SAVE_HOTEL_ID',
    SAVE_HOTEL_LIST = '[Hotel] SAVE_HOTEL_LIST',
    SAVE_RATING_COUNT = '[Hotel] SAVE_RATING_COUNT',
    SAVE_PREFECTURE_VALUE = '[Hotel] SAVE_PREFECTURE_VALUE',
    FILTER_HOTELS_CONTENT = '[Hotel] FILTER_HOTELS_CONTENT',
    FILTER_HOTELS_SEARCH_BOX = '[Hotel] FILTER_HOTELS_SEARCH_BOX'
}